requirements.txt
matplotlib==3.7.1
numpy==1.22.3
pandas==1.5.3
Pillow==9.4.0
Pillow==10.4.0
scikit_learn==1.3.0
torch==1.12.1+cu102
torchvision==0.13.1+cu102
